Converted edit billable service ooverlay to wirkspace
