+++ packages/esm-billing-app/src/root.scss
@@ -3,7 +3,6 @@
 @use '@carbon/colors';
 @import '~@openmrs/esm-styleguide/src/vars';
 
-
 .productiveHeading01 {
   @include type.type-style('heading-01');
 }
