+++ packages/esm-billing-app/src/routes.json
@@ -183,6 +183,12 @@
       "component": "editPurchasePriceForm",
       "title": "Edit Purchase Price Form",
       "type": "other-form"
+    },
+    {
+      "name": "edit-billable-service-form",
+      "component": "editBillableServiceForm",
+      "title": "Edit Billable Service Form",
+      "type": "other-form"
     }
   ]
 }
